+++ Hemologica-DAO/src/main/java/org/hemologica/dao/model/ResultsCode.java
@@ -22,7 +22,7 @@ public class ResultsCode {
 	@Id
 	@GeneratedValue(strategy=GenerationType.AUTO)
 	@Column(name="results_code_id")
-	private int resultsCodeId;
+	private Long resultsCodeId;
 
 	@Column(name="results_code_label")
 	private String resultsCodeLabel;
@@ -38,11 +38,11 @@ public class ResultsCode {
 	@JoinColumn(name="results_code_snomed")
 	private Concept concept;
 
-	public int getResultsCodeId() {
+	public Long getResultsCodeId() {
 		return resultsCodeId;
 	}
 
-	public void setResultsCodeId(int resultsCodeId) {
+	public void setResultsCodeId(Long resultsCodeId) {
 		this.resultsCodeId = resultsCodeId;
 	}
